chore: migrations eslint fixes
