+++ lib/types/lists.ts
@@ -1,3 +1,5 @@
+import { z } from "zod";
+
 import { WorkType } from "./generated/db";
 import { DetailedMovieData } from "./movie";
 
@@ -29,3 +31,12 @@ export type DetailedWorkListItem<T = DetailedMovieData> = WorkListItem &
 
 export type DetailedReviewListItem<T = DetailedMovieData> = ReviewListItem &
   ExternalWorkData<T>;
+
+export const listInsertDtoSchema = z.object({
+  type: z.nativeEnum(WorkType),
+  title: z.string(),
+  externalId: z.string().optional(),
+  imageUrl: z.string().optional(),
+});
+
+export type ListInsertDto = z.infer<typeof listInsertDtoSchema>;
