+++ netlify/functions/repositories/WorkRepository.ts
@@ -1,6 +1,6 @@
 import { hasValue } from "../../../lib/checks/checks.js";
 import { WorkType } from "../../../lib/types/generated/db.js";
-import { ListInsertDto } from "../../../lib/types/ListDto.js";
+import { ListInsertDto } from "../../../lib/types/lists.js";
 import { db } from "../utils/database";
 import { getDetailedWorks } from "../utils/tmdb";
